Remove logic from inside `iocSpecBuilder` for storing actions.
With the action facet in place it will do this work, allowing us to remove this code from the builder.
Also some minor cleanup where we can use `Q` as a function instead of `Q.when`.
The bootstrapped now references actions of the container instead of the spec, which is much nicer.